Reduced memory allocated for temporary buffer. It need not be more than the size of the ringbuffer itself

+++ BoomAndroid/openslplayer/src/main/jni/bufferprovider/include/RingBuffer.h
@@ -41,7 +41,7 @@ namespace gdpl {
         int             _writePtr;
         int             _writeBytesAvail;
 
-        int16_t*        _tempBuffer;
+        uint8_t*        _tempBuffer;
     };
 };
